+++ packages/preview-middleware-client/test/__mock__/sap/ui/fl/apply/_internal/flexObjects/FlexObjectFactory.ts
@@ -6,14 +6,18 @@ interface ChangeContent {
 }
 
 export default {
-    createFromFileContent: jest.fn().mockImplementation((oFileContent: {selector: object, changeType: string, layer: string}): FlexChange<ChangeContent> => {
-        return {
-            getSelector: jest.fn().mockReturnValue(oFileContent.selector),
-            getChangeType: jest.fn().mockReturnValue(oFileContent.changeType),
-            getLayer: jest.fn().mockReturnValue(oFileContent.layer),
-            getDefinition: jest.fn(),
-            getContent: jest.fn(),
-            setContent: jest.fn()
-        }
-    })
-}
+    createFromFileContent: jest
+        .fn()
+        .mockImplementation(
+            (oFileContent: { selector: object; changeType: string; layer: string }): FlexChange<ChangeContent> => {
+                return {
+                    getSelector: jest.fn().mockReturnValue(oFileContent.selector),
+                    getChangeType: jest.fn().mockReturnValue(oFileContent.changeType),
+                    getLayer: jest.fn().mockReturnValue(oFileContent.layer),
+                    getDefinition: jest.fn(),
+                    getContent: jest.fn(),
+                    setContent: jest.fn()
+                };
+            }
+        )
+};